requirements.txt
django==5.2.6
djangorestframework==3.16.1
environs==14.3.0